+++ src/utils/helpers/features.ts
@@ -27,3 +27,5 @@ export const supportsTagsInPatch = supportsShortUrlTitle;
 export const supportsBotVisits = serverMatchesVersions({ minVersion: '2.7.0' });
 
 export const supportsCrawlableVisits = supportsBotVisits;
+
+export const supportsQrErrorCorrection = serverMatchesVersions({ minVersion: '2.8.0' });